+++ js/app/app.js
@@ -3,7 +3,6 @@ var
   Router = require( "react-router" ),
   routes = require("./routes");
 
-
 Router.run( routes, function(Handler) {
   React.render(
     <Handler/>,