+++ js/app/stores/LoaderStore.js
@@ -7,7 +7,8 @@ var
 
   ActionTypes = KanbanerConstants.ActionTypes,
 
-  LOADER_CHANGE = 'loader-change',
+  LOADER_STORE_INIT = "loader-store-init",
+  LOADER_CHANGE = "loader-change",
 
   isShowLoader = false,
 
@@ -25,6 +26,13 @@ var
   hideLoader = function() {
     isShowLoader = false;
     LoaderStore.emitLoaderChange();
+  },
+
+  /**
+   * init loader store
+   */
+  initLoaderStore = function() {
+    LoaderStore.emitLoaderStoreInit()
   },
 
   LoaderStore = assign({}, EventEmitter.prototype, {
@@ -56,8 +64,30 @@ var
      * Emit loader change
      */
     emitLoaderChange: function() {
-      console.log("[LoaderStore] Send loader event");
       this.emit(LOADER_CHANGE);
+    },
+
+    /**
+     * Add loader store init event listener
+     * @param callback
+     */
+    addLoaderStoreInitListener: function(callback) {
+      this.on(LOADER_STORE_INIT, callback)
+    },
+
+    /**
+     * Remove loader store init event listener
+     * @param callback
+     */
+    removeLoaderStoreInitListener: function(callback) {
+      this.on(LOADER_STORE_INIT, callback);
+    },
+
+    /**
+     * send Loader store init event
+     */
+    emitLoaderStoreInit: function() {
+      this.emit(LOADER_STORE_INIT);
     }
   });
 
@@ -65,6 +95,9 @@ LoaderStore.dispatchToken = KanbanerDispatcher.register(function( payload) {
   var action = payload.action;
 
   switch( action.type ) {
+    case ActionTypes.APP_INIT:
+      initLoaderStore();
+      break;
     case ActionTypes.LOADER_SHOW:
       showLoader();
       break;