+++ js/app/stores/UserStore.js
@@ -7,11 +7,12 @@ var
 
   ActionTypes = KanbanerConstants.ActionTypes,
 
+  USER_STORE_INIT = "user-store-init",
   USER_LOGGED_IN = "user-logged-in",
   USER_LOGGED_OUT = "user-logged-out",
   USER_AUTHENTICATION_FAILED = "user-authentication-failed",
 
-  user = null,
+  user = null;
   token = null;
 
 /**
@@ -22,6 +23,7 @@ var
 var loginUser = function( authUser, authToken ) {
   user = authUser;
   token = authToken;
+
   UserStore.emitUserLoggedIn();
 };
 
@@ -35,7 +37,12 @@ var sendLoginFailed = function() {
 var logoutUser = function() {
   user = null;
   token = null;
+
   UserStore.emitUserLoggedOut();
+};
+
+var initUserStore = function() {
+  UserStore.emitUserStoreInit();
 };
 
 var UserStore = assign({}, EventEmitter.prototype, {
@@ -49,9 +56,18 @@ var UserStore = assign({}, EventEmitter.prototype, {
 
   /**
    * Returns user
+   * @returns {object}
    */
   getUser: function() {
     return user;
+  },
+
+  /**
+   * Returns token
+   * @returns {null|*}
+   */
+  getToken: function() {
+    return token;
   },
 
   /**
@@ -121,14 +137,38 @@ var UserStore = assign({}, EventEmitter.prototype, {
    */
   emitAuthenticationFailed: function() {
     this.emit(USER_AUTHENTICATION_FAILED);
+  },
+
+  /**
+   * Add user store init event listener
+   * @param {function} callback
+   */
+  addUserStoreInitListener: function (callback) {
+    this.on(USER_STORE_INIT, callback);
+  },
+
+  /**
+   * Remove user store init event listener
+   */
+  removeUserStoreInitListener: function(callback) {
+    this.removeListener(USER_STORE_INIT, callback);
+  },
+
+  /**
+   * Send user store init event
+   */
+  emitUserStoreInit: function() {
+    this.emit(USER_STORE_INIT)
   }
 });
 
 UserStore.dispatcherToken = KanbanerDispatcher.register(function( payload ) {
   var action = payload.action;
 
-  console.log( action.type );
   switch( action.type ) {
+    case ActionTypes.APP_INIT:
+      initUserStore();
+      break;
     case ActionTypes.USER_LOGIN_SUCCESS:
       loginUser( action.user, action.token );
       break;